refactor : repository 인터페이스 추상화
