+++ src/main/java/woowa/cafe/repository/QuestionRepository.java
@@ -1,36 +1,14 @@
 package woowa.cafe.repository;
 
 import woowa.cafe.domain.Question;
-import woowa.frame.core.annotation.Component;
 
-import java.lang.reflect.Field;
 import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
 
-@Component
-public class QuestionRepository {
+public interface QuestionRepository {
 
-    private Map<String, Question> database = new ConcurrentHashMap<>();
+    void save(Question question);
 
-    public void save(Question question) {
-        try {
-            String uuid = UUID.randomUUID().toString();
-            Field id = Question.class.getDeclaredField("id");
-            id.setAccessible(true);
-            id.set(question, uuid);
-            database.put(question.getId(), question);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new RuntimeException(e);
-        }
-    }
+    List<Question> findAll();
 
-    public List<Question> findAll() {
-        return database.values().stream().toList();
-    }
-
-    public Question findById(String id) {
-        return database.get(id);
-    }
+    Question findById(String id);
 }
